refactor(//core/partitioning): Reorganizing partitioning deps
Reorganizing the partitioning dependencies so that there is a clearer
relationship between major compiler modules
