+++ core/partitioning/partitioning.h
@@ -2,8 +2,8 @@
 
 #include <vector>
 
-#include "core/conversion/conversion.h"
-#include "core/conversion/evaluators/eval_util.h"
+#include "core/ir/ir.h"
+#include "core/partitioning/PartitionInfo.h"
 #include "core/partitioning/SegmentedBlock.h"
 #include "core/util/prelude.h"
 #include "torch/csrc/jit/ir/ir.h"
@@ -12,10 +12,12 @@ namespace trtorch {
 namespace core {
 namespace partitioning {
 
+typedef std::vector<SegmentedBlock> PartitionedGraph;
+
 std::vector<SegmentedBlock> Partition(
     std::shared_ptr<torch::jit::Graph> g,
-    std::vector<conversion::InputRange>& input_ranges,
-    const conversion::TorchFallback& fallback_info);
+    std::vector<ir::InputRange>& input_ranges,
+    const PartitionInfo& partition_info);
 
 } // namespace partitioning
 } // namespace core
