+++ core/partitioning/shape_analysis.h
@@ -1,10 +1,12 @@
-#include "SegmentedBlock.h"
+#include "core/partitioning/SegmentedBlock.h"
+#include "core/ir/ir.h"
+
 
 namespace trtorch {
 namespace core {
 namespace partitioning {
 
-std::vector<torch::jit::IValue> generateRandomInputs(std::vector<conversion::InputRange>& input_ranges);
+std::vector<torch::jit::IValue> generateRandomInputs(std::vector<ir::InputRange>& input_ranges);
 
 void getSegmentsOutputByRunning(
     SegmentedBlock& seg_block,
